Rework the whole provider for v0.4.0 (now requires Predis v0.8).
Nothing changes for single-client configurations, but there is
a BC for multiple-clients configurations since the addition of
the `boot()` mechanism in Silex essentially made impossible to
keep this feature alive without artificially boot()-ing the
provider (which was no good).

Multiple clients are now supported only when using the new
`Predis\Silex\MultiPredisServiceProvider` but the way clients
are configured remained the same.

Clients in such configuration are accessible in the following
manner:

  $client = $app['predis'][$alias];

It is no more possible to configure a default client responding
to `$app['predis']` due to the nature of the above change.