[tests] Remove silex from require-dev dependencies.
PredisServiceProvider can actually be used without Silex as it depends
on Pimple now after all, examples will not work out of the box though.
In the future we could even rename the namespace from Predis\Silex to
Predis\Pimple to reflect this shift, but it is best to wait to see how
Silex 2.0 will evolve in the near future.